% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameliacombinr.R
\name{ameliacombinr}
\alias{ameliacombinr}
\title{Combine multiple imputations for a list of regression models}
\usage{
ameliacombinr(modellist, aout, subsets = as.list(rep(NA, length(modellist))))
}
\arguments{
\item{modellist}{A list of regression models, entered as formulas}

\item{aout}{The output from the amelia() function}

\item{subsets}{A list of numeric or logical vectors, one for each model, for running the model on a subset of the data}
}
\description{
This function lets you input a list of regression models and retrieve coefficients and standard errors based on a combination of imputed data sets.
}
\examples{
mtcarsmiss<-mtcars
mtcarsmiss$wt[c(5,10,15)]<-NA #create missing values
require(Amelia)
mtcarsmiss_im<-amelia(mtcarsmiss) #run multiple imputations
m1f<-as.formula(mpg~wt) #formula for model 1
m2f<-as.formula(mpg~wt+cyl) #formula for model 2
require(hjorthr)
#ameliacombinr(modellist=list(m1f,m2f),aout=mtcarsmiss_im)
}
\keyword{amelia}
\keyword{imputation}
\keyword{regression}

