% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginsplotdf.R
\name{marginsplotdf}
\alias{marginsplotdf}
\title{Function for creating data for marginsplot}
\usage{
marginsplotdf(model, xterm, zterm, zseq)
}
\description{
This function lets you easily create a plot of marginal effects from a regression interaction model..
}
\details{
@param model A regression model object, e.g. model1
 @param xterm The term for which to plot effects on y, as character
 @param zterm The term along which to plot effects of x on y
 @param zseq Sequence of values of z for which to plot the marginal effect of x
 @examples
 m1<-lm(mpg~cyl*wt,data=mtcars)
 require(hjorthr)
 mdf<-marginsplotdf(m1,"wt","cyl",seq(from=min(mtcars$cyl),to=max(mtcars$cyl),by=.5))

require(ggplot2)
 ggplot(mdf,aes(x=z,y=dydx,ymin=dydx-1.96*se,ymax=dydx+1.96*se)) +
   geom_line() +
   geom_ribbon(alpha=.2)
}

